Reversed connection. still not working though
